feat(water): v1.8.3 Meld Water Update — vanilla-MC bed + shore + wetland
Underwater bed picker with independent per-block domain-warped value noise →
organic vanilla-MC-style patches (CLAY/SAND/DIRT/COARSE_DIRT) on GRAVEL
background. Rare 5-13 cell MAGMA + SOUL_SAND vents at depth >= 5 produce
bubble columns in-game. Slope curve sqrt(dist/span). Width-aware dunes
amplitude 2-4 with domain-warped sampling for prominent waves. STONE bedrock
under all sea cells + 12-cell under-fill kills AIR pockets.

Underwater fauna mix: short SEAGRASS + TALL_SEAGRASS + SEA_PICKLE; KELP with
min 3 cells, variable 30-100 % column height. Veg now plants ABOVE the dune
top so it never carves holes into the bed surface (fixes empty-whitelist
always-replace bug).

Shore palette: smooth value_noise static fade SAND -> COARSE_DIRT ->
GRASS_BLOCK over 5-cell band. Stoney shore (STONE_BRICKS etc neighbour)
gets COBBLE 50% + MOSSY_COBBLE 30% + COARSE_DIRT 20%. Shore-land rare
cattail + sugar cane near water-edge.

Wetland: per-subtype branch matching arnis-source-water structure with
post-pass MOSS_BLOCK 8-ring + COARSE_DIRT 2-ring around water puddles.
Tiered cattail = 1 or 2 stalks + single BROWN_CANDLE block with
candles=1/2/3/4 NBT state.

Plumbing:
* 9 new Block IDs (256-265, u16 widening): MAGMA_BLOCK, SUGAR_CANE, KELP,
  TALL_SEAGRASS_BOTTOM/TOP, SEA_PICKLE, BROWN_CANDLE_2/3/4, SOUL_SAND.
* --seed (alias --tile-invariant-rendering) ties global NOISE_SEED via
  OnceLock so identical seed reproduces identical bed/dune/shore patterns.
* sweep_floating_veg post-pass removes cattails, grass, candles, sugar
  cane and flowers placed over water cells or roads. Trees (LOG/LEAVES)
  explicitly excluded.
* 5x5 bridge ring in OSM water carve matches LC_WATER pass.

Cargo version 1.8.3 (Meld fork numbering, downstream of Teddy 2.8.1 and
louis-e v2.8.0 base).

+++ src/models_3d/wikidata/placement.rs
@@ -982,7 +982,7 @@ mod tests {
         assert_eq!(material_to_rgb("nonsense"), None);
     }
 
-    fn block_ids(palette: &[Block]) -> Vec<u8> {
+    fn block_ids(palette: &[Block]) -> Vec<u16> {
         palette.iter().map(|b| b.id()).collect()
     }
